pipeline/requirements.txt
functions-framework==3.5.0
google-cloud-firestore==2.16.0
httpx==0.27.0
